+++ scene/realistic-light-and-shadows/README.metadata.json
@@ -1,34 +1,35 @@
 {
-    "category": "Scene",
-    "description": "Show realistic lighting and shadows for the time of day.",
-    "ignore": false,
-    "images": [
-        "RealisticLightingAndShadows.png"
-    ],
-    "keywords": [
-        "3D",
-        "lighting",
-        "realism",
-        "realistic",
-        "rendering",
-        "shadows",
-        "sun",
-        "time",
-        "ArcGISScene",
-        "LightingMode",
-        "SceneView.setSunLighting",
-        "SceneView.setSunTime"
-    ],
-    "redirect_from": "",
-    "relevant_apis": [
-        "ArcGISScene",
-        "LightingMode",
-        "SceneView.setSunLighting",
-        "SceneView.setSunTime"
-
-    ],
-    "snippets": [
-        "src/main/java/com.esri.samples.realistic_light_and_shadows/RealisticLightAndShadows.java"
-    ],
-    "title": "Realistic Light and Shadows Sample"
+  "category": "Scene",
+  "description": "Show realistic lighting and shadows for the time of day.",
+  "ignore": false,
+  "images": [
+    "RealisticLightingAndShadows.png"
+  ],
+  "keywords": [
+    "3D",
+    "lighting",
+    "realism",
+    "realistic",
+    "rendering",
+    "shadows",
+    "sun",
+    "time",
+    "ArcGISScene",
+    "LightingMode",
+    "SceneView.setSunLighting",
+    "SceneView.setSunTime"
+  ],
+  "redirect_from": "",
+  "relevant_apis": [
+    "ArcGISScene",
+    "LightingMode",
+    "SceneView.setSunLighting",
+    "SceneView.setSunTime"
+  ],
+  "snippets": [
+    "src/main/java/com.esri.samples.realistic_lighting_and_shadows/RealisticLightAndShadowsSample.java",
+    "src/main/java/com.esri.samples.realistic_lighting_and_shadows/RealisticLightAndShadowsController.java",
+    "src/main/resources/realistic_light_and_shadows/main.fxml"
+  ],
+  "title": "Realistic Lighting and Shadows"
 }
