+++ scene/realistic-light-and-shadows/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'Realistic Light and Shadows'
+rootProject.name = 'Realistic Lighting and Shadows'
