--- scene/realistic-light-and-shadows/src/main/java/com/esri/samples/realistic_light_and_shadows/RealisticLightAndShadowsLauncher.java
@@ -1,11 +1,0 @@
-package com.esri.samples.realistic_light_and_shadows;
-
-/**
- * Wrapper required for launching a JavaFX 11 app through Gradle or from a jar.
- */
-public class RealisticLightAndShadowsLauncher {
-
-  public static void main(String[] args) {
-    RealisticLightAndShadowsSample.main(args);
-  }
-}
