+++ scene/realistic-light-and-shadows/src/main/java/com/esri/samples/realistic_lighting_and_shadows/RealisticLightingAndShadowsLauncher.java
@@ -1,0 +1,11 @@
+package com.esri.samples.realistic_lighting_and_shadows;
+
+/**
+ * Wrapper required for launching a JavaFX 11 app through Gradle or from a jar.
+ */
+public class RealisticLightingAndShadowsLauncher {
+
+  public static void main(String[] args) {
+    RealisticLightingAndShadowsSample.main(args);
+  }
+}
